fix: remove warnings

+++ edit_distance_hashing.h
@@ -5,8 +5,8 @@
 template <typename Seq>
 int EditDistanceHashParallel(const Seq &a, const Seq &b) {
   // build sparse table
-  size_t n = a.size();
-  size_t m = b.size();
+  int n = a.size();
+  int m = b.size();
   if (n == 0) {
     return m;
   }
